Moved the rest of the functional tests over too
