--- Spinit.CosmosDb.FunctionalTests/Properties/AssemblyInfo.cs
@@ -1,3 +1,0 @@
-﻿using Xunit;
-
-[assembly: CollectionBehavior(DisableTestParallelization = true)]
